// src/app/shared/model/response/customer-response.model.ts
export class CustomerResponseModel {
  CustomerID: number;
  CompanyID: number;
  GroupCustomerID: number;
  SaleID: number;
  TypeCustomerID: number;
  StoreID: number;
  Code: string;
  TaxCode: string;
  Password: string;
  NameCustomer: string;
  Avatar: string;
  VAT: number;
  Phone: string;
  Address: string;
  Description: string;
  Position: number;
  Birthday: Date;
  Status: number;
  UserCreated: string;
  UserUpdated: string;
  DateCreated: Date;
  DataUpdated: Date;
}